reduce time taken for elastic.cpp (static) tests to run
- moved source file to top of CMakeLists.txt list to give it head start
- broke repetetive tests into 8 source files for parallel compilation
- results:
  - Clang: before=18.4s, moving elastic.cpp up=14.8s, 8-way=12.3
  - GCC:   before=22.7s, moving elastic.cpp up:20.5s, 8-way=12.7
