+++ src/test/elastic_common.h
@@ -4,6 +4,9 @@
 //  (See accompanying file ../../LICENSE_1_0.txt or copy at
 //          http://www.boost.org/LICENSE_1_0.txt)
 
+// like fixed_point_common.h this file contains tests which can be
+// performed on different types from different .cpp files.
+
 #include <elastic.h>
 
 #include <tuple>
@@ -512,71 +515,4 @@ struct elastic_test_with_integer_digits :
 ////////////////////////////////////////////////////////////////////////////////
 // trigger elastic tests against a range of values for IntegerDigits parameter
 
-template
-struct elastic_test_with_integer_digits<0>;
-
-template
-struct elastic_test_with_integer_digits<1>;
-
-template
-struct elastic_test_with_integer_digits<-1>;
-
-template
-struct elastic_test_with_integer_digits<2>;
-
-template
-struct elastic_test_with_integer_digits<-2>;
-
-template
-struct elastic_test_with_integer_digits<5>;
-
-template
-struct elastic_test_with_integer_digits<-5>;
-
-template
-struct elastic_test_with_integer_digits<6>;
-
-template
-struct elastic_test_with_integer_digits<-6>;
-
-template
-struct elastic_test_with_integer_digits<8>;
-
-template
-struct elastic_test_with_integer_digits<-8>;
-
-template
-struct elastic_test_with_integer_digits<13>;
-
-template
-struct elastic_test_with_integer_digits<-13>;
-
-template
-struct elastic_test_with_integer_digits<16>;
-
-template
-struct elastic_test_with_integer_digits<-16>;
-
-template
-struct elastic_test_with_integer_digits<19>;
-
-template
-struct elastic_test_with_integer_digits<-19>;
-
-template
-struct elastic_test_with_integer_digits<27>;
-
-template
-struct elastic_test_with_integer_digits<-27>;
-
-template
-struct elastic_test_with_integer_digits<31>;
-
-template
-struct elastic_test_with_integer_digits<-31>;
-
-template
-struct elastic_test_with_integer_digits<43>;
-
-template
-struct elastic_test_with_integer_digits<-43>;
+// (from elastic?.cpp source files)
